wt-mini-project/backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
